refactor(components): split layout components
