+++ src/routes/_terminal.settings.about.index.tsx
@@ -2,7 +2,7 @@ import { createFileRoute } from "@tanstack/react-router"
 import { GitFork, type Info, ScrollText, ShieldCheck } from "lucide-react"
 import { type ComponentProps, useMemo } from "react"
 import { FadeHeader } from "@/components/fade-header.tsx"
-import { VerticalNav } from "@/components/vertial-nav.tsx"
+import { VerticalNav } from "@/components/vertical-nav.tsx"
 import { useTranslation } from "@/hooks/use-translation.ts"
 import type { TranslationKey } from "@/i18n/resources.ts"
