+++ src/routes/restore-account.tsx
@@ -2,7 +2,7 @@ import { createFileRoute, useNavigate } from "@tanstack/react-router"
 import { LoaderCircleIcon } from "lucide-react"
 import { useEffect } from "react"
 
-import { PhoneViewport } from "@/components/skeleton.tsx"
+import { PhoneViewport } from "@/components/phone-viewport.tsx"
 import { settingsQuery } from "@/core/modules/app-settings/app-settings-queries.ts"
 import { useEvoluQuery } from "@/hooks/use-evolu-query.ts"
 import { useTranslation } from "@/hooks/use-translation.ts"
